app/facturekiller_v3/requirements.txt
# FactureKiller V3 - Dépendances

# Framework Web
Flask==3.0.0
flask-cors==4.0.0

# OCR et traitement d'images
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.1.0
pillow-heif==0.13.1

# Traitement de données
pandas==2.1.4
openpyxl==3.1.2
numpy==1.26.2

# IA et analyse
anthropic==0.7.7
python-dotenv==1.0.0

# Utilitaires
python-dateutil==2.8.2
requests==2.31.0 